% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_speed.R
\name{change_speed}
\alias{change_speed}
\title{Change the playback speed of an audio signal}
\usage{
change_speed(input_signal, rate)
}
\arguments{
\item{input_signal}{numeric}

\item{rate}{numeric, desired rate of change to the speed.
To increase the speed, pass in a value greater than 1.0.
To decrease the speed, pass in a value between 0.0 and 1.0.}
}
\value{
numeric, vector representing the audio signal with changed speed.
}
\description{
Change the playback speed of an audio signal
}
