% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_pass_filter.R
\name{low_pass_filter}
\alias{low_pass_filter}
\title{Filter audio signal, attenuating frequencies below a cut off.}
\usage{
low_pass_filter(input_signal, cutoff_frequency)
}
\arguments{
\item{input_signal}{numeric}

\item{cutoff_frequency}{numeric, cutoff frequency of the low pass filter.}
}
\value{
numeric, vector representing the audio signal with frequencies below the cut off attenuated by 24 db.
}
\description{
Filter audio signal, attenuating frequencies below a cut off.
}
